pyproject.toml
[project]
name = "primordial-orchestrator"
version = "0.1.0"
description = "Dynamic orchestrator with agent discovery and delegation"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "Andy Browning" }]

dependencies = [
    "deepagents>=0.2",
    "langgraph>=0.2",
    "langchain>=0.3",
    "langchain-anthropic",
    "python-dotenv",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"
